server/requirements.txt
# Kafka
confluent-kafka>=2.6.0

# GCP / BigQuery
google-cloud-bigquery>=3.25.0

# API (FastAPI)
fastapi>=0.115.0
uvicorn>=0.32.0

# Utilities
python-dotenv>=1.0.0
